+++ home/views.py
@@ -2,6 +2,7 @@ from django.conf import settings
 from django.shortcuts import *
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.contrib.auth.models import User
+from django.views.decorators.http import require_POST
 
 from social.apps.django_app.default.models import UserSocialAuth
 
@@ -116,6 +117,8 @@ def beats(request):
     return render_to_response('beats.html', context, context_instance=RequestContext(request))
 
 from django.contrib.auth import logout as auth_logout
+@login_required
+@require_POST
 def logout(request):
     """Logs out user"""